libretro: Add gearboy_libretro.info
These info files are required for proper operation of the libretro
core. It is customary to host these in the libretro-super repo
to make them easy to maintain, but given that the libretro code
was merged in this repo and this repo also maintain release tags
it would be useful to have the gearboy_libretro.info file provided
in those release tarballs for distro packages. I suspect this file
will not need to be updated often, but if it does the downside is that
it will need to be updated in two repos.